backend/requirements.txt
# FastAPI 核心
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# 数据库
sqlalchemy==2.0.25
asyncpg==0.29.0
greenlet==3.0.3

# Supabase
supabase==2.3.4
python-jose[cryptography]==3.3.0

# Pydantic
pydantic==2.6.1
pydantic-settings==2.1.0

# HTTP 客户端 (使用 supabase 兼容的版本)
httpx>=0.24,<0.26
openai==1.12.0

# 工具
python-dotenv==1.0.1
